Homepage Javascript and CSS Tweaks
Tried changing some of the CSS to make the different areas on the
homepage stand out from one another. I also added javascript/jquery to
animate the navigation menu on the left and the chatbox from the bottom.
Lastly I did a bunch of small tweaks and changes to improve usability.